+++ CHANGES.md
@@ -8,16 +8,27 @@ Version 0.2.0
 
 To be released.
 
+ -  Added [`areValuesEqual()`][] function.
+ -  Added [`areDictionariesEqual()`][] function.
+
+[`areValuesEqual()`]:
+https://deno.land/x/bencodex@0.2.0/mod.ts?s=areValuesEqual
+[`areDictionariesEqual()`]:
+https://deno.land/x/bencodex@0.2.0/mod.ts?s=areDictionariesEqual
+
 
 Version 0.1.1
 -------------
 
 Released on March 3, 2023.
 
- -  `BencodexDictionary` and `RecordView` now shows easier inspection
-     representions on the eyes both on Deno ([`Deno.inspect()`]) and
-     Node.js ([`util.inspect()`]).
+ -  [`BencodexDictionary`] and [`RecordView`] now shows easier inspection
+    representions on the eyes both on Deno ([`Deno.inspect()`]) and
+    Node.js ([`util.inspect()`]).
 
+
+[`BencodexDictionary`]: https://deno.land/x/bencodex@0.1.1/mod.ts?s=BencodexDictionary
+[`RecordView`]: https://deno.land/x/bencodex@0.1.1/mod.ts?s=RecordView
 [`Deno.inspect()`]: https://deno.land/api?s=Deno.inspect
 [`util.inspect()`]: https://nodejs.org/api/util.html#utilinspectobject-options